rebase: fit to new project structure + lint, mypy
